requirements-build.txt
# Requirements for building Windows executable
# Core dependencies from pyproject.toml
matplotlib>=3.10.5
pandas>=2.3.1
scikit-learn>=1.7.1
PyQt6>=6.5.0
olefile>=0.47
scipy
numpy

# Build tools
pyinstaller>=6.0.0

# Additional dependencies that might be needed
pillow  # For image handling in matplotlib and icon conversion